provide and trigger product cubit
