+++ lib/core/service/service_locator.dart
@@ -6,19 +6,26 @@ import 'package:fruits_hup/modules/auth/domain/repos/auth_repo.dart';
 import '../database/cache/cache_helper.dart';
 import 'package:get_it/get_it.dart';
 
+import '../repos/products_repos/product_repo.dart';
+import '../repos/products_repos/products_repos_implementation.dart';
 import 'service_firebase_firestore.dart';
 
 final getIt = GetIt.instance;
 
 void setUpServiceLocator() {
+  //! Shared Preferences
   getIt.registerSingleton<CacheHelper>(CacheHelper());
-  getIt.registerSingleton<FirebaseAuthService>(FirebaseAuthService());
-    getIt.registerSingleton<DatabaseService>(FirabaseFirestoreService());
 
+  //! Authentication
+  getIt.registerSingleton<FirebaseAuthService>(FirebaseAuthService());
+  getIt.registerSingleton<DatabaseService>(FirabaseFirestoreService());
   getIt.registerSingleton<AuthRepo>(
     AuthRepoImplement(
       firebaseAuthService: getIt<FirebaseAuthService>(),
-      databaseService:  getIt<DatabaseService>()
+      databaseService: getIt<DatabaseService>(),
     ),
   );
+
+  //! Home
+  getIt.registerSingleton<ProductRepo>(ProductReposImplementation(databaseService: getIt<DatabaseService>()));
 }
