+++ lib/modules/home/presentation/widgets/custom_home_view_body.dart
@@ -1,0 +1,32 @@
+
+import 'package:flutter/material.dart';
+
+import '../../../../core/widgets/custom_best_celling_grid_view_widget.dart';
+import '../../../../core/widgets/custom_search_bar_widget.dart';
+import 'custom_best_celling_header_text_widget.dart';
+import 'custom_home_app_bar_widget.dart';
+import 'custom_home_featured_list_widget.dart';
+
+class HomeViewBody extends StatelessWidget {
+  const HomeViewBody({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomScrollView(
+      physics: const BouncingScrollPhysics(),
+      slivers: [
+        const SliverToBoxAdapter(child: SizedBox(height: 30)),
+        const SliverToBoxAdapter(child: CustomHomeAppBarWidget()),
+        const SliverToBoxAdapter(child: SizedBox(height: 12)),
+        const SliverToBoxAdapter(child: CustomSearchBarWidget()),
+        const SliverToBoxAdapter(child: SizedBox(height: 12)),
+        const SliverToBoxAdapter(child: CustomHomeFeaturedListWidget()),
+        const SliverToBoxAdapter(child: SizedBox(height: 12)),
+        SliverToBoxAdapter(child: BestCellingHeaderText()),
+        const SliverToBoxAdapter(child: SizedBox(height: 8)),
+        BestCellingGridView(),
+        const SliverToBoxAdapter(child: SizedBox(height: 24)),
+      ],
+    );
+  }
+}
